add example customize
