+++ examples/cpp/CMakeLists.txt
@@ -22,12 +22,6 @@ target_link_libraries (simple_hgraph_sq8 vsag_static)
 add_executable (example_conjugate_graph example_conjugate_graph.cpp)
 target_link_libraries(example_conjugate_graph vsag)
 
-add_executable (custom_memory_allocator custom_memory_allocator.cpp)
-target_link_libraries(custom_memory_allocator vsag)
-
-add_executable (custom_logger custom_logger.cpp)
-target_link_libraries(custom_logger vsag)
-
 add_executable (vsag_ext_example vsag_ext_example.cpp)
 target_link_libraries(vsag_ext_example vsag)
