+++ examples/cpp/todo_examples/203_custom_thread_pool.cpp
@@ -12,3 +12,73 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
+#include <iostream>
+
+#include "default_thread_pool.h"
+#include "safe_thread_pool.h"
+#include "vsag/logger.h"
+#include "vsag/thread_pool.h"
+#include "vsag/vsag.h"
+
+int
+main() {
+    vsag::Options::Instance().logger()->SetLevel(vsag::Logger::kINFO);
+
+    /******************* Customize Thread Pool *****************/
+    auto pool = std::make_shared<vsag::SafeThreadPool>(new vsag::DefaultThreadPool(16), true);
+    vsag::Resource resource(nullptr, pool.get());
+    vsag::Engine engine(&resource);
+
+    vsag::init();
+
+    /******************* Prepare Base Dataset *****************/
+    int64_t num_vectors = 1000;
+    int64_t dim = 128;
+    auto ids = new int64_t[num_vectors];
+    auto vectors = new float[dim * num_vectors];
+
+    std::mt19937 rng;
+    rng.seed(47);
+    std::uniform_real_distribution<> distrib_real;
+    for (int64_t i = 0; i < num_vectors; ++i) {
+        ids[i] = i;
+    }
+    for (int64_t i = 0; i < dim * num_vectors; ++i) {
+        vectors[i] = distrib_real(rng);
+    }
+    auto base = vsag::Dataset::Make();
+    base->NumElements(num_vectors)->Dim(dim)->Ids(ids)->Float32Vectors(vectors);
+
+    /******************* Create DiskANN Index *****************/
+    auto diskann_build_paramesters = R"(
+    {
+        "dtype": "float32",
+        "metric_type": "l2",
+        "dim": 128,
+        "diskann": {
+            "max_degree": 16,
+            "ef_construction": 200,
+            "pq_sample_rate": 0.5,
+            "pq_dims": 9,
+            "use_pq_search": true,
+            "use_async_io": false,
+            "use_bsa": true
+        }
+    }
+    )";
+    auto index = engine.CreateIndex("diskann", diskann_build_paramesters).value();
+
+    /******************* Build DiskANN Index *****************/
+    if (auto build_result = index->Build(base); build_result.has_value()) {
+        std::cout << "After Build(), Index DiskANN contains: " << index->GetNumElements()
+                  << std::endl;
+    } else if (build_result.error().type == vsag::ErrorType::INTERNAL_ERROR) {
+        std::cerr << "Failed to build index: internalError" << std::endl;
+        exit(-1);
+    }
+
+    engine.Shutdown();
+
+    return 0;
+}
