+++ examples/cpp/todo_examples/CMakeLists.txt
@@ -13,6 +13,15 @@ target_link_libraries(104_index_fresh_hnsw vsag)
 add_executable (105_index_brute_force 105_index_brute_force.cpp)
 target_link_libraries (105_index_brute_force vsag_static)
 
+add_executable (201_custom_allocator 201_custom_allocator.cpp)
+target_link_libraries (201_custom_allocator vsag)
+
+add_executable (202_custom_logger 202_custom_logger.cpp)
+target_link_libraries (202_custom_logger vsag)
+
+add_executable (203_custom_thread_pool 203_custom_thread_pool.cpp)
+target_link_libraries (203_custom_thread_pool vsag)
+
 add_executable (301_feature_filter 301_feature_filter.cpp)
 target_link_libraries (301_feature_filter vsag_static)
